requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
cachetools==6.2.6
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
dill==0.4.0
faiss-cpu==1.13.2
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
narwhals==2.17.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
opentelemetry-proto==1.39.1
packaging==26.0
pillow==12.1.1
preshed==3.0.12
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
#pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart_open==7.5.1
smmap==5.0.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.47
sqlparse==0.5.5
srsly==2.5.2
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
fastapi==0.115.8
python-multipart==0.0.20